Added ViewModel. Renamed namespaces.
